Types avro (#3)
* Corrected NumberType#doubleValue to return doubleValue and not floatValue

* First commit of the types-avro module

* Added clarification that `BaseBigDecimalTypeConversion` maps to Avro's primitive type `double` and not logical-type `decimal`, which comes at the risk of loosing precision in the number representation in the conversion between the concrete `BigDecimalType`'s internal `BigDecimal` and the corresponding `double` value in Avro!

Improved Java doc and examples

* Updated copyright to cover the year 2022

* Improved Conversion validation error messages
Added test for array, map and union fields using LogicalTypes

+++ types/src/main/java/dk/cloudcreate/essentials/types/NumberType.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2021-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -111,9 +111,8 @@ public abstract class NumberType<NUMBER_TYPE extends Number, CONCRETE_TYPE exten
 
     @Override
     public double doubleValue() {
-        return value.floatValue();
+        return value.doubleValue();
     }
-
 
     @Override
     public int hashCode() {
